split MQTT into common and device objects
