+++ common/transport/mqtt/.npmignore
@@ -1,0 +1,6 @@
+coverage/
+devdoc/
+test/
+samples/
+src/
+lib/**/*.map
