+++ common/transport/mqtt/devdoc/mqtt_base_requirements.md
