+++ common/transport/mqtt/devdoc/mqtt_translate_errors_requirements.md
