+++ common/transport/mqtt/index.d.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export { MqttBase } from './lib/mqtt_base';
+export { translateError } from './lib/mqtt_translate_error';
