+++ common/transport/mqtt/package.json
@@ -1,27 +1,24 @@
 {
-  "name": "azure-iot-device-mqtt",
-  "version": "1.1.18",
-  "description": "MQTT transport for Azure IoT device SDK",
+  "name": "azure-iot-mqtt-base",
+  "version": "2.0.0",
+  "description": "MQTT operations used by Azure IoT device and service SDKs",
   "author": "Microsoft Corporation",
   "license": "MIT",
-  "main": "index.js",
+  "main": "main.js",
   "typings": "index.d.ts",
   "dependencies": {
     "azure-iot-common": "1.1.12",
-    "azure-iot-device": "1.1.18",
-    "azure-iot-mqtt-base": "2.0.0",
-    "debug": "^3.0.1",
-    "machina": "^2.0.0"
+    "machina": "^2.0.0",
+    "mqtt": "^2.13.0"
   },
   "devDependencies": {
-    "@types/mqtt": "0.0.34",
-    "@types/node": "^8.0.30",
+    "@types/node": "^8.0.28",
     "chai": "^4.1.2",
     "es5-shim": "^4.5.9",
     "istanbul": "^0.4.5",
     "jshint": "^2.9.4",
     "mocha": "^3.2.0",
-    "sinon": "^3.3.0",
+    "sinon": "^3.2.1",
     "tslint": "^5.1.0",
     "typescript": "2.5.2"
   },
@@ -33,8 +30,8 @@
     "unittest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test.js",
     "alltest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test*.js",
     "ci": "npm -s run lint && npm -s run build && npm -s run alltest-min && npm -s run check-cover",
-    "test": "npm -s run lint && npm -s run build && npm -s run unittest",
-    "check-cover": "istanbul check-coverage --statements 95 --branches 81 --functions 92 --lines 96"
+    "test": "npm -s run lint && npm -s run build && npm -s run alltest && npm -s run check-cover",
+    "check-cover": "istanbul check-coverage --statements 92 --branches 80 --functions 83 --lines 94"
   },
   "engines": {
     "node": ">= 0.10"
