+++ common/transport/mqtt/readme.md
@@ -1,0 +1,6 @@
+#azure-iot-mqtt-base
+Internal library of AMQP-specific components shared between the Azure IoT device and service SDKs.
+
+[![npm version](https://badge.fury.io/js/azure-iot-mqtt-base.svg)](https://badge.fury.io/js/azure-iot-mqtt-base)
+
+NOTE: You generally don't want to reference this package directly. The MQTT components you need should be exposed by packages in the device (`azure-iot-device`) or service (`azure-iothub`) SDKs.
