+++ common/transport/mqtt/src/mqtt_translate_error.ts
@@ -6,7 +6,7 @@
 import { errors } from 'azure-iot-common';
 
 /**
- * @method          module:azure-iot-device-mqtt.translateError
+ * @method          module:azure-iot-mqtt-base.translateError
  * @description     Convert an error returned by MQTT.js into a transport-agnistic error.
  *
  * @param {Object}        mqttError the error returned by the MQTT.js library
