+++ common/transport/mqtt/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+        "outDir": "./lib",
+        "target":"es5",
+        "sourceMap": true,
+        "declaration": true
+    },
+    "include": [
+        "./src/**/*.ts"
+    ],
+    "exclude": [
+        "./lib"
+    ]
+}
